// src/utils/enums/Login.enum.ts
export enum MenuPathEnum {
  SETTINGS = 'CONFIG',
  SETTINGS_HOME = 'CONFIGURACIONES',
  LOGOUT = 'CERRAR SESIóN',
  LOGIN = 'Login',
  JORNADA = 'JORNADAS',
  JORNADA_HOME = 'JORNADA',
  TAREA_HOME = 'TAREAS',
  TAREA = 'Tarea',
  TAREA_CREATE = 'CrearTarea',
  TAREA_EDIT = 'EditarTarea',
  TAREA_DETALLE = 'DetalleTarea',
  TAREA_ATTEND = 'Tarea',
  CLIENTE_HOME = 'CLIENTES',
  CLIENTE = 'ClienteList',
  CLIENTE_DETAIL = 'ClienteDetalle',
  CLIENTE_CREATE = 'ClienteCrear',
}